reverse-geocode: simplify with helper LookupLatLon
In usages I've found myself having to make a raw query ie
```go
&mapbox.ReverseGeocodeRequest{Query: fmt.Sprintf("%v,%v", lat, lon)}
```

This is a very awkard usage and in fact the users of the API
shouldn't have to worry about which comes first: lat or lon.

This change simplifies it by introducing a helper method
LookupLatLon which takes in (lat, lon) and performs a reverse
geocode lookup as was with the original ReverseGeocoding request.
It also simplifies the example usage.